fix: stabilize runtime MCP test tempdir cleanup on Windows CI
